feature: separate profile screen content in separated files.
